clean up package
