+++ backend/ibft/verifier.go
@@ -84,7 +84,7 @@ func (i *backendIBFT) IsValidSender(msg *protoIBFT.Message) bool {
 	}
 
 	//	verify the sender is in the active validator set
-	if !i.currentValidatorSet.Includes(validatorAddress) {
+	if !i.activeValidatorSet.Includes(validatorAddress) {
 		//i.logger.Error("committed seal generated by validator not in the current set")
 
 		return false
@@ -101,7 +101,7 @@ func (i *backendIBFT) IsProposer(id []byte, height, round uint64) bool {
 		return false
 	}
 
-	nextProposer := i.currentValidatorSet.CalcProposer(round, extractProposer(previousHeader))
+	nextProposer := i.activeValidatorSet.CalcProposer(round, extractProposer(previousHeader))
 	if !bytes.Equal(nextProposer.Bytes(), id) {
 		return false
 	}
@@ -141,7 +141,7 @@ func (i *backendIBFT) IsValidCommittedSeal(proposalHash, seal []byte) bool {
 		return false
 	}
 
-	if !i.currentValidatorSet.Includes(validator) {
+	if !i.activeValidatorSet.Includes(validator) {
 		i.logger.Error("seal generated by validator not in the active set", "validator", validator.String())
 
 		return false
